src/main/java/com/it_academy/practice/junit_basics/Calculator.java
package com.it_academy.practice.junit_basics;

public class Calculator {

    private int a;
    private int b;

    public Calculator(int a, int b) {
        this.a = a;
        this.b = b;
    }

    public float getA() {
        return a;
    }

    public void setA(int a) {
        this.a = a;
    }

    public float getB() {
        return b;
    }

    public void setB(int b) {
        this.b = b;
    }

    public float calculate(char operation) {
        switch (operation) {
            default: {
                return 0;
            }
            case '-': {
                return a - b;
            }
            case '+': {
                return a + b;
            }
            case '/': {
                try {
                    return a / b;
                } catch (ArithmeticException e) {
                    System.out.println("Division by zero is not allowed");
                }
            }
            case '*': {
                return a * b;
            }
            case '^': {
                return (float) Math.pow(a, b);
            }
            case '√': {
                return (float) Math.pow(a, 1.0 / b);
            }
        }
    }


}
